+++ src/components/common/actions-panel/ActionsPanel.module.css
