+++ src/components/common/left-panel/LeftPanel.js
