+++ src/components/common/left-panel/LeftPanel.module.css
