+++ src/components/common/widgets/CategoryWidget.js
