+++ src/components/views/Kpis/Kpis.test.js
