+++ src/routes.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Home from './pages/Home/Home';
-import Kpis from './pages/Kpis/Kpis';
+import Home from './components/views/Home/Home';
+import Kpis from './components/views/Kpis/Kpis';
 
 const routes = [
   { path: '/', element: <Home /> },
